Testing API priority NextJS
I originally had an issue with loading priority on the LBL website.

The main issue was, the Hero video would take priority before any API calls were made or displayed.

I had the idea to bundle the video's source into the API call the page needs to make so both the video AND the api data would be fetched at the same time.

This should mean no more waiting for video to fully download before it displays the API data!

(I'm fighting god and I just got a good right hook in there)